requirements.txt
pylint
pytest
requests
ipykernel
pandas
python-dotenv
pyodbc
aiohttp
asyncio